+++ fk_vtr.h
@@ -4,12 +4,12 @@
 #include  <strings.h>
 
 typedef struct {
-    uint32_t init_len;
-    uint32_t len;
+    size_t init_len;
+    size_t len;
     void **array;
 } fk_vtr_t;
 
-fk_vtr_t *fk_vtr_create(uint32_t init_len);
+fk_vtr_t *fk_vtr_create(size_t init_len);
 void fk_vtr_destroy(fk_vtr_t *vtr);
 
 #define fk_vtr_len(vtr)		((vtr)->len)
